Azure AI PDF Analyzer/requirements.txt
azure-ai-vision-imageanalysis
streamlit